requirements.txt
streamlit
pandas
datetime
plotly
reportlab
unidecode
psycopg2-binary
